Refactor to make PMD and Tests happy
